backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
pydantic==2.9.2
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.12
python-dotenv==1.0.1
apscheduler==3.10.4
aiosmtplib==3.0.2
email-validator==2.2.0
Pillow==10.4.0
slowapi==0.1.9
httpx==0.27.2
